ExpanderTab had only one use outside its package - moved
